mail

+++ main.py
@@ -2,6 +2,7 @@ from email import message
 import email
 from json import decoder
 from flask import Flask, render_template, request, redirect, url_for, session
+from requests.sessions import Session
 from flask_mysqldb import MySQL
 import MySQLdb.cursors
 import re
@@ -19,6 +20,8 @@ import sys
 import jwt
 import requests
 import json
+
+
 from time import time
 import datetime
 #sys.path.insert(0, 'Rportal/config')
@@ -266,13 +269,14 @@ def security():
 @app.route('/R-Portal/add_staff', methods=['GET','POST'])
 def staff():
     msg = ''
-    if request.method == 'POST' and 'staff_name' in request.form and 'staff_username' in request.form and 'staff_password' in request.form:
+    if request.method == 'POST' and 'staff_name' in request.form and 'staff_username' in request.form and 'staff_password' in request.form  and 'post' in request.form:
         
         staff_username = request.form['staff_username']
         staff_mobile = request.form['staff_mobile']
         passtext  = request.form['staff_password']
         staff_password  = hashlib.sha256((passtext).encode('utf-8')).hexdigest()
         staff_name = request.form['staff_name']
+        post = request.form['post']
         cursor = mysql.connection.cursor(MySQLdb.cursors.DictCursor)
         cursor.execute('SELECT * FROM staff WHERE staff_username = %s AND staff_mobile = %s', (staff_username, staff_mobile))
         account = cursor.fetchone()
@@ -283,7 +287,7 @@ def staff():
         elif not staff_username or not staff_password or not staff_mobile:
             msg = 'Please fill out the form!'
         else:
-            cursor.execute('INSERT INTO staff VALUES (NULL, %s, %s, %s, %s, %s, DEFAULT)', (staff_username , staff_password , staff_name , staff_mobile, session['code']))
+            cursor.execute('INSERT INTO staff VALUES (NULL, %s, %s, %s, %s, %s , DEFAULT, %s )', (staff_username , staff_password , staff_name , staff_mobile, session['code'],post))
             mysql.connection.commit()
             msg = 'You have successfully registered!'
     elif request.method == 'POST':
@@ -586,17 +590,17 @@ def logout():
 def people():
     if 'secretary' in session:
         cursor = mysql.connection.cursor(MySQLdb.cursors.DictCursor)
-        cursor.execute('SELECT Mname, member_status, Mmobile ,Mid, Mflatno,Mwing from secretary inner join member on secretary.Scode = member.Mcode WHERE Sid = %s AND member_status=%s', (session['id'],'active'))
+        cursor.execute('SELECT Mname, member_status, Mmobile ,Mid, Mflatno,Mwing ,Memail from secretary inner join member on secretary.Scode = member.Mcode WHERE Sid = %s AND member_status=%s', (session['id'],'active'))
         account = cursor.fetchall()     
         cursor1 = mysql.connection.cursor(MySQLdb.cursors.DictCursor)
-        cursor1.execute('SELECT Sname,secretarty_status, Smobile from secretary WHERE Sid = %s AND secretarty_status=%s', (session['id'],'active'))
+        cursor1.execute('SELECT Sname,secretarty_status, Smobile ,Semail from secretary WHERE Sid = %s AND secretarty_status=%s', (session['id'],'active'))
         account1 = cursor1.fetchone()  
         cursor2 = mysql.connection.cursor(MySQLdb.cursors.DictCursor)
         cursor2.execute('SELECT   security_id,security_name, security_status, security_mobile from secretary inner join security on secretary.Scode = security.security_code WHERE Sid = %s AND security_status=%s', (session['id'],'active'))
         account2 = cursor2.fetchall()     
         cursor3 = mysql.connection.cursor(MySQLdb.cursors.DictCursor)
         cursor3.execute('SELECT staff_name, staff_status, staff_mobile ,staff_id from secretary inner join staff on secretary.Scode = staff.staff_code WHERE Sid = %s AND staff_status=%s', (session['id'],'active'))
-        account3 = cursor3.fetchall()        
+        account3 = cursor3.fetchall()         
         return render_template('secretary/people.html', account=account, account1=account1, account2=account2, account3=account3)
     else:
         return logout()
@@ -605,7 +609,7 @@ def people():
 def inpeople():
     if 'secretary' in session:
         cursor = mysql.connection.cursor(MySQLdb.cursors.DictCursor)
-        cursor.execute('SELECT Mname, member_status, Mmobile, Mid from secretary inner join member on secretary.Scode = member.Mcode WHERE Sid = %s AND member_status=%s', (session['id'],'inactive'))
+        cursor.execute('SELECT Mname, member_status, Mmobile, Mid ,Memail ,Mflatno,Mwing from secretary inner join member on secretary.Scode = member.Mcode WHERE Sid = %s AND member_status=%s', (session['id'],'inactive'))
         account = cursor.fetchall()     
         cursor1 = mysql.connection.cursor(MySQLdb.cursors.DictCursor)
         cursor1.execute('SELECT Sname,secretarty_status, Smobile from secretary WHERE Sid = %s AND secretarty_status=%s', (session['id'],'inactive'))
@@ -626,17 +630,24 @@ def allow_members():
         cursor = mysql.connection.cursor(MySQLdb.cursors.DictCursor)
         cursor.execute('SELECT Mid, Mname, member_status, Mmobile, Mwing, Mflatno, Memail, Sname, name from secretary inner join member inner join society on secretary.Scode = member.Mcode AND secretary.Scode = society.code WHERE Sid = %s AND member_status=%s', (session['id'],'Request'))
         account = cursor.fetchall() 
+
         return render_template('secretary/allow_members.html', account=account )
     else:
         return logout()
     
-@app.route('/R-Portal/a_members/<int:Mid>')
+@app.route('/R-Portal/a_members/<int:Mid>' , methods=['GET', 'POST'])
 def a_members(Mid):
     if 'secretary' in session:
-        cursor = mysql.connection.cursor(MySQLdb.cursors.DictCursor)
-        cursor.execute("UPDATE member SET member_status = %s WHERE Mid = %s" ,('active',Mid,)) 
-        mysql.connection.commit()
-        msg = 'Member Allowed'
+        if request.method == 'POST' and 'Memail' in request.form and 'Mname':
+            cursor = mysql.connection.cursor(MySQLdb.cursors.DictCursor)
+            cursor.execute("UPDATE member SET member_status = %s WHERE Mid = %s" ,('active',Mid,)) 
+            mysql.connection.commit()
+            email = request.form['Memail']
+            Mname = request.form['Mname']
+            msg = Message('Your Joining Request is accepted' ,sender ='Rportal<me@Rportal.com', recipients = [email])
+            msg.body ="Hello!" + Mname + " \n your account activate for login please login on rportal ." + part4
+            mail.send(msg) 
+
         return allow_members()
     else:
         return logout()
@@ -644,21 +655,32 @@ def a_members(Mid):
 @app.route('/R-Portal/i_members/<int:Mid>', methods=['GET', 'POST'])
 def i_members(Mid):
     if 'secretary' in session:
-        cursor = mysql.connection.cursor(MySQLdb.cursors.DictCursor)
-        cursor.execute("UPDATE member SET member_status = %s WHERE Mid = %s" ,('inactive',Mid,)) 
-        mysql.connection.commit()
-        msg = 'Member Allowed'
-        return people()
+        if request.method == 'POST' and 'Memail' in request.form and 'Mname':
+            cursor = mysql.connection.cursor(MySQLdb.cursors.DictCursor)
+            cursor.execute("UPDATE member SET member_status = %s WHERE Mid = %s" ,('inactive',Mid,)) 
+            mysql.connection.commit()
+            email = request.form['Memail']
+            Mname = request.form['Mname']
+            msg = Message('Account Inactive' ,sender ='Rportal<me@Rportal.com', recipients = [email])
+            msg.body ="Hello!" + Mname + " \n You have Temporary Inactive from secreatry please contact with your secretary." + part4
+            mail.send(msg)  
+            return people()
     else:
         return logout()
 
 @app.route('/R-Portal/ac_members/<int:Mid>', methods=['GET', 'POST'])
 def ac_members(Mid):
     if 'secretary' in session:
-        cursor = mysql.connection.cursor(MySQLdb.cursors.DictCursor)
-        cursor.execute("UPDATE member SET member_status = %s WHERE Mid = %s" ,('active',Mid,)) 
-        mysql.connection.commit()
-        msg = 'Member Allowed'
+        if request.method == 'POST' and 'Memail' in request.form and 'Mname':
+            cursor = mysql.connection.cursor(MySQLdb.cursors.DictCursor)
+            cursor.execute("UPDATE member SET member_status = %s WHERE Mid = %s" ,('active',Mid,)) 
+            mysql.connection.commit()
+            email = request.form['Memail']
+            Mname = request.form['Mname']
+            msg = Message('Account activated' ,sender ='Rportal<me@Rportal.com', recipients = [email])
+            msg.body ="Hello!" + Mname + " \n You have active from secreatry any query please contact with your secretary." + part4
+            mail.send(msg)  
+        
         return inpeople()
     else:
         return logout()
@@ -683,13 +705,18 @@ def r_members():
     else:
         return logout()
 
-@app.route('/R-Portal/d_members/<int:Mid>')
+@app.route('/R-Portal/d_members/<int:Mid>' , methods=['GET', 'POST'])
 def d_members(Mid):
     if 'secretary' in session:
-        cursor = mysql.connection.cursor(MySQLdb.cursors.DictCursor)
-        cursor.execute('DELETE FROM member WHERE Mid = %s',[Mid]) 
-        mysql.connection.commit()
-        msg = 'Member Delete'
+        if request.method == 'POST' and 'Memail' in request.form and 'Mname':
+            cursor = mysql.connection.cursor(MySQLdb.cursors.DictCursor)
+            cursor.execute('DELETE FROM member WHERE Mid = %s',[Mid]) 
+            mysql.connection.commit()
+            email = request.form['Memail']
+            Mname = request.form['Mname']
+            msg = Message('Account Delete' ,sender ='Rportal<me@Rportal.com', recipients = [email])
+            msg.body ="Hello!" + Mname + " \n Your account has deleted from secreatry please contact with your secretary.\n" + part4
+            mail.send(msg)  
         return people()
     else:
         return logout()
@@ -819,7 +846,7 @@ def createnotice():
 def complaint():
     if 'secretary' in session:
         cursor = mysql.connection.cursor(MySQLdb.cursors.DictCursor)        
-        cursor.execute('SELECT * FROM complaint WHERE complaint_code = %s', (session['code'],))
+        cursor.execute('SELECT * FROM complaint WHERE complaint_code = %s ', (session['code'],))
         account = cursor.fetchall()
         return render_template('secretary/complaint.html', account=account)
     else:
@@ -839,19 +866,20 @@ def createmeeting():
             agenda = request.form['agenda']
             date = request.form['date']
             password = request.form['password']
-
             time_now = datetime.datetime.now()
             exprieation_time = time_now + datetime.timedelta(seconds=20)
             rounded_off_exp_time = round(exprieation_time.timestamp())
 
             headers ={"alg" : "HS256", "type": "JWT"}
-            payload ={"iss" : "DVKW9UngTg2lI1FnqrQVWA" ,"exp" : rounded_off_exp_time}
+          
+            payload ={"iss" : "DVKW9UngTg2lI1FnqrQVWA" ,"exp" : rounded_off_exp_time  }
             encode_jwt = jwt.encode(payload,"ZIN4AsaniiKYBBA0cQHLSupJOIZwMEdbcRm2" , algorithm="HS256")
-            email= "aashutoshmali1460@gmail.com"
+            email =  "aashutoshmali1460@gmail.com"
             url= "https://api.zoom.us/v2/users/{}/meetings".format(email)
-            obj = {"topic" : topic ,"start_time" : date + '\t' + start_time , "duration" : duration , "password" : password  }
+           # date=datetime.datetime(2021,11,23,23,50).strfttime("%Y-%m-%dT%H:%M:%SZ")
+            obj = {"topic" : "test" ,"start_time" :"2021,11,23,23,50"  , "duration" : 30 , "password" : "12345"  }
             header = {"authorization": "Bearer {}".format(encode_jwt) }
-            create_meeting = requests.post (url,json=obj ,headers = header) 
+            create_meeting = requests.post(url,json=obj ,headers = header) 
        
             print( create_meeting.text)
             msg = create_meeting.text
@@ -895,16 +923,27 @@ def viewnotice():
 def member_complaint():
     if 'member' in session:
         msg = ''
-        if request.method == 'POST' and 'complaint_subject' in request.form and 'complaint_message' in request.form:
+        cursor1 = mysql.connection.cursor(MySQLdb.cursors.DictCursor)
+        cursor1.execute('SELECT post from staff WHERE staff_code= %s AND staff_status=%s', (session['code'],'active',))
+        account = cursor1.fetchall() 
+        cursor2 = mysql.connection.cursor(MySQLdb.cursors.DictCursor)
+        cursor2.execute('SELECT * from complaint WHERE complaint_username= %s ', (session['username'],))
+        account2 = cursor2.fetchall() 
+        if request.method == 'POST' and 'complaint_subject' in request.form and 'complaint_message' in request.form and 'complaint_against'in request.form :  
             compaint_subject = request.form['complaint_subject']
             complaint_message = request.form['complaint_message']
+            complaint_against = request.form['complaint_against']
             cursor = mysql.connection.cursor(MySQLdb.cursors.DictCursor)
-            cursor.execute('INSERT INTO complaint VALUES (NULL, %s, %s, %s, %s, %s, DEFAULT)', (session['username'] , session['name'] , compaint_subject, complaint_message, session['code']))
+            cursor.execute('INSERT INTO complaint VALUES (NULL, %s, %s, %s, %s, %s, DEFAULT,%s,NULL , NULL)', (session['username'] , session['name'] , compaint_subject, complaint_message, session['code'],complaint_against))
             mysql.connection.commit()
+           
             msg = 'Complaint Added Succsesfully!'
+            
         elif request.method == 'POST':
             msg = 'Please fill out the form!'
-        return render_template('member/member_complaint.html', msg=msg)
+          
+           
+        return render_template('member/member_complaint.html', msg=msg , account=account,account2= account2)
     else:
         return logout()
